move sssom config to its own file, add match enum and config
