+++ src/koza/model/config/source_config.py
@@ -9,16 +9,14 @@ import zipfile
 from dataclasses import field
 from enum import Enum
 from pathlib import Path
-from typing import Dict, List, Union, Literal
+from typing import Dict, List, Union
 import yaml
 
-from loguru import logger
 from pydantic import StrictFloat, StrictInt, StrictStr
 from pydantic.dataclasses import dataclass
-from sssom.parsers import parse_sssom_table
-from sssom.util import filter_prefixes, merge_msdf
 
 from koza.model.config.pydantic_config import PydanticConfig
+from koza.model.config.sssom_config import SSSOMConfig
 
 
 class FilterCode(str, Enum):
@@ -130,129 +128,6 @@ class DatasetDescription:
     provided_by: str = None
     license: str = None
     rights: str = None
-
-
-@dataclass()
-class SSSOMConfig:
-    """SSSOM config options
-
-    Attributes:
-        files: List of SSSOM files to use
-        filter_prefixes: Optional list of prefixes to filter by
-        subject_target_prefixes: Optional list of prefixes to use for subject mapping
-        object_target_prefixes: Optional list of prefixes to use for object mapping
-
-    """
-
-    files: List[Union[str, Path]] = field(default_factory=list)
-    filter_prefixes: List[str] = field(default_factory=list)
-    subject_target_prefixes: List[str] = field(default_factory=list)
-    object_target_prefixes: List[str] = field(default_factory=list)
-
-    predicates = {"exact": ["skos:exactMatch"], "narrow": ["skos:narrowMatch"], "broad": ["skos:broadMatch"]}
-
-    def __post_init_post_parse__(self):
-        logger.debug("Building SSSOM Dataframe...")
-        self.df = self._merge_and_filter_sssom()
-        logger.debug("Building SSSOM Lookup Table...")
-        self.lut = self._build_sssom_lut()
-
-    def apply_mapping(self, entity: dict) -> dict:
-        """Apply SSSOM mappings to an edge record."""
-
-        if self._has_mapping(entity["subject"], self.subject_target_prefixes):
-            entity["original_subject"] = entity["subject"]
-            entity["subject"] = self._get_mapping(entity["subject"], self.subject_target_prefixes)
-
-        if self._has_mapping(entity["object"], self.object_target_prefixes):
-            entity["original_object"] = entity["object"]
-            entity["object"] = self._get_mapping(entity["object"], self.object_target_prefixes)
-
-        return entity
-
-    def _merge_and_filter_sssom(self):
-        mapping_sets = []
-        for file in self.files:
-            msdf = parse_sssom_table(file)
-            mapping_sets.append(msdf)
-        new_msdf = merge_msdf(*mapping_sets)
-        filters = (self.subject_target_prefixes + self.object_target_prefixes) + list(
-            set(self.filter_prefixes) - set(self.subject_target_prefixes) - set(self.object_target_prefixes)
-        )
-        logger.debug(f"Filtering SSSOM by {filters}")
-        new_msdf = filter_prefixes(
-            df=new_msdf.df, filter_prefixes=filters, require_all_prefixes=False, features=new_msdf.df.columns
-        )
-
-        return new_msdf
-
-    def _build_sssom_lut(self) -> Dict:
-        """Build a lookup table from SSSOM mapping dataframe."""
-        sssom_lut = {}
-        for _, row in self.df.iterrows():
-            subject_id = row["subject_id"]
-            object_id = row["object_id"]
-            predicate = row["predicate_id"]
-            # Add exact match mappings in both directions
-            sssom_lut = self._set_mapping(subject_id, object_id, predicate, match="exact", lut=sssom_lut)
-            sssom_lut = self._set_mapping(object_id, subject_id, predicate, match="exact", lut=sssom_lut)
-            # TODO add narrow and broad match mappings
-            # sssom_lut = self._set_mapping(subject_id, object_id, predicate, match='broad', lut=sssom_lut)
-            # sssom_lut = self._set_mapping(object_id, subject_id, predicate, match='narrow', lut=sssom_lut)
-        return sssom_lut
-
-    def _has_match(self, predicate, match: Literal["exact", "narrow", "broad"]):
-        """Check if a predicate has a match."""
-        if match == "exact":
-            return predicate in self.predicates["exact"]
-        if match == "narrow":
-            # return predicate in self.predicates["narrow"] or predicate in self.predicates["exact"]
-            logger.warning("Narrow match not yet implemented")
-            return False
-        if match == "broad":
-            # return predicate in self.predicates["broad"] or predicate in self.predicates["exact"]
-            logger.warning("Broad match not yet implemented")
-            return False
-
-    def _has_mapping(self, id, target_prefixes=None):
-        """Check if an ID has a mapping."""
-        if target_prefixes is None:
-            return id in self.lut
-        else:
-            if id not in self.lut:
-                return False
-            for target_prefix in target_prefixes:
-                if target_prefix in self.lut[id]:
-                    return True
-            logger.warning(f"No mapping found for {id} in {target_prefixes}: {self.lut[id]}")
-            return False # No mapping found
-
-    def _get_mapping(self, id, target_prefixes):
-        """Get the mapping for an ID."""
-        for target_prefix in target_prefixes:
-            if target_prefix in self.lut[id]:
-                return self.lut[id][target_prefix]
-        raise KeyError(f"Could not find mapping for {id} in {target_prefixes}: {self.lut[id]}")
-
-    def _set_mapping(
-        self, original_id, mapped_id, predicate, match: Literal["exact", "narrow", "broad"], lut: Dict[str, dict]
-    ):
-        """Set a mapping for an ID."""
-        original_prefix = original_id.split(":")[0]
-        mapped_prefix = mapped_id.split(":")[0]
-        target_prefixes = self.subject_target_prefixes + self.object_target_prefixes
-        if (
-            original_prefix in self.filter_prefixes or len(self.filter_prefixes) == 0
-        ) and mapped_prefix in target_prefixes:
-            if original_id not in lut:
-                lut[original_id] = {}
-            if mapped_prefix in lut[original_id]:
-                logger.warning(f"Duplicate mapping for {original_id} to {mapped_prefix}")
-            elif self._has_match(predicate, match):
-                lut[original_id][mapped_prefix] = mapped_id
-            else:
-                pass  # do something else
-        return lut
 
 
 @dataclass(config=PydanticConfig)
